+++ pystac/extensions/label.py
@@ -8,7 +8,7 @@ from pystac.serialization.identify import STACJSONDescription, STACVersionID
 from pystac.extensions.hooks import ExtensionHooks
 from pystac.utils import StringEnum, get_required, map_opt
 
-SCHEMA_URI = "https://stac-extensions.github.io/label/v1.0.0/schema.json"
+SCHEMA_URI = "https://stac-extensions.github.io/label/v1.0.1/schema.json"
 
 PREFIX = "label:"
 
@@ -788,7 +788,10 @@ class SummariesLabelExtension(SummariesExtension):
 
 class LabelExtensionHooks(ExtensionHooks):
     schema_uri: str = SCHEMA_URI
-    prev_extension_ids = {"label"}
+    prev_extension_ids = {
+        "label",
+        "https://stac-extensions.github.io/label/v1.0.0/schema.json",
+    }
     stac_object_types = {pystac.STACObjectType.ITEM}
 
     def get_object_links(